MOD: Separate methods of the Database.java class into separate classes
