+++ src/database/DatabaseShowListController.java
@@ -1,0 +1,9 @@
+package database;
+
+public class DatabaseShowListController {
+//	Methods:
+	// Show lists []->{ )
+	public void showList() {
+		// TODO: Show list feature!
+	}
+}
