+++ src/main/MainApplication.java
@@ -1,17 +1,12 @@
 package main;
 
 import view.WelcomeView;
-
-import database.DatabaseAddRemoveController;
-import database.Database;
-import models.Course;
-import models.Student;
-import models.Teacher;
+import database.AddDefaultObjects;
 
 public class MainApplication {
 	public static void main(String[] args) {
 		// Create default objects for test:
-		new Database().addDefaultUserCards();
+		new AddDefaultObjects().add();
 		
 		// Start program:
 		new WelcomeView().start();
@@ -20,8 +15,8 @@ public class MainApplication {
 
 /*
 * TODO:
-* 1. Optimize and remove static methods
-* 2. Optimize and remove infinite loops
+* 1. Optimize and remove static methods ---
+* 2. Optimize and remove infinite loops ---
 * 3. Add feature with course: search, show list and create
 * 		(Show who enrolled in every course by request)
 *
